Initial commit, dotfiles project.
The idea is that I can keep several of the files I use with some
frequency in a single place. I already have a dotfiles project
somewhere else in my computer, but sadly that project is very messy.
From now on I should be able to keep this project under wraps and better
organized here.